[UP] rename utils.py to misc.py
