+++ storch/__init__.py
@@ -1,3 +1,3 @@
 
-from storch.utils import *
+from storch.misc import *
 from storch.version import __version__
