+++ storch/misc.py
